# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Second Life Kids API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkout/create-session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create Stripe checkout session for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionBody"
      responses:
        "200":
          description: Stripe checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /checkout/quote-request:
    post:
      operationId: submitQuoteRequest
      tags: [checkout]
      summary: Submit XL quote request (no payment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequestBody"
      responses:
        "201":
          description: Quote request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error
...
    CreateCheckoutSessionBody:
      type: object
      properties:
        bookingId:
          type: string
      required:
        - bookingId
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
        sessionId:
          type: string
      required:
        - url
        - sessionId
    QuoteRequestBody:
      type: object
      properties:
        customerName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        suburb:
          type: string
        postcode:
          type: string
        description:
          type: string
        preferredDate:
          type: string
          nullable: true
      required:
        - customerName
        - email
        - phone
        - suburb
        - postcode
        - description
    QuoteRequest:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        customerName:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [PENDING, CONTACTED, CLOSED]
      required:
        - id
        - createdAt
        - customerName
        - email
        - status
...